Rerandomize shuffle seeds when replicating ShuffleDataset.
If the graph and shuffle seeds are both set to 0, our convention is to behave non-deterministically by choosing random seeds.

Before this change, the chosen seeds would be baked into the ShuffleDataset so that if we convert the ShuffleDataset to a graphdef and back, the seeds would keep whatever random values were chosen in the original ShuffleDataset. However, this probably isn't desired since replicated datasets will have identical seeds, but setting seeds to (0, 0) typically means the user wants true randomness.

This CL addresses the issue by remembering the originally configured seed values and using those when converting a dataset to a graphdef.

The added test fails without the changes to ShuffleDataset.

PiperOrigin-RevId: 284640034
Change-Id: Ia36612cb7ca1e8848c63602fff72f166da43002e