Add the numeric verification support in the quantization pass
when `-tfl-numeric-verify` flag is set, for each op being quantized, the
floating-point op is preserved in the model.  Both the results from the
floating-point op and the quantized op are sent to the numeric verify op for
verifying with an error tolerance, which is set by the `-tfl-error-tolerance`
flag.

If any operands of this floating-point op are constant, the floating-point
constants will be used directly.

If the `-tfl-single-layer-verify` is enable, this result of this floating-point
op will not be used by the next layer, so the errors are not accumulated. If it
is disabled, the floating-point results are used by the floating-point ops at
the next layer, so the error will be accumulated.

When the flag is enabled, the produced model should be valid.

PiperOrigin-RevId: 285418911
Change-Id: I5ef519ee18f00799136a1c845483f0d71c3e0c00