Allow partial padding_values definitions with 'None' to indicate default values
For example, a nested structure like ([], [], {'hello': [], 'world': []}) could
have a padding_values provided like (0, 1, None) to indicate that the third
tuple element should have its nested structure be default-padded.
This makes it easier to specify the padding of only one of the values, when the
rest is fine with defaults.

PiperOrigin-RevId: 286453825
Change-Id: I1aae75bd3c25b55175751504f79d116c3f9ee267