Fix non-deterministic shuffling when reshuffle_each_iteration=False
Before this fix, if no shuffling seeds were set, we would reshuffle on each iteration even when `reshuffle_each_iteration=False`. This bug was introduced when `AsGraphDef` was changed in cl/284640034 so that if seeds are set non-deterministically, serializing and deserializing via `AsGraphDef` would re-randomize the seeds. This triggered re-randomization because `AsGraphDef` is used by `OptimizeDatasetOp` each iteration to rewrite the dataset graph.

To fix the issue, we add a flag to `SerializationContext` so that the serialization performed for rewriting doesn't trigger rerandomization.

PiperOrigin-RevId: 286626730
Change-Id: Ie37e61d1059e01fabb0b7945418357d9f05a269f