[MLIR:TF/XLA]  Create no-op in empty island when converting to control dialect.
When converted to control dialect during export, an empty island is replaced
with all the control inputs in the nodes depending on it. This might create
quadratic number of edges in certain cases (N control inputs and N nodes
depending on it). So we create a no-op to group the input control edges.

PiperOrigin-RevId: 287873185
Change-Id: Ic8219b319e864becac595e211f162bdc05ae773e