Add the passes to support default range quantization
To test a quantized model, the user might want to specify a pair of default
min/max values for any tensors which don't have sufficient quantization
parameters (FakeQuant ops were not placed in the right location). This new pass
runs after the normal quantization passes, and tries to use the default range
to quantize the rest of the model.

PiperOrigin-RevId: 288804672
Change-Id: I2e00cc3a17cc5e582e1c3f7cd643c199d6060459