Uses `tf_cc_shared_object` instead of `cc_library`.
Windows filesystem plugin fails to work if built with `cc_library`. Hence, we have to use `cc_binary` or the `tf_cc_libary` macro that TF provides. The macro is better as it provides more control and will be easier to adapt in the future (this is useful as now we need `linkstatic = False` for this to work but in the near future we might have Bazel support for shared libraries, in which case things would be simpler).

Furthermore, switching to the macro allows us to clean the `BUILD` file, separating the private targets from the public one. Also, removing the TODOs as now things are fixed.

Part of the work for modular filesystem plugins. For more details, consult the RFC at https://github.com/tensorflow/community/blob/master/rfcs/20190506-filesystem-plugin-modular-tensorflow.md

PiperOrigin-RevId: 288811849
Change-Id: Iefc8315be69de46098e1502f15b46e6bb5fafaae