Unify the handling of basic and composite symbols by using nonlocal instead of explicit function arguments and return values in the functions that represent control flow blocks. This change also adds more uniform dtype/shape checking across different types of for-loop iteration.
In essence, this changes the representation of control flow from:

```
def loop_body(<loop_vars>)
  ...
  return <new_loop_vars>
```

to:

```
def loop_body():
  nonlocal <loop_vars>
  ...
```

Requires Python 3.

This representation matches side effects more accurately, in particular as when the original control flow block modifies variables that are closed over.
It also simplifies the code by using a common mechanism to handle basic and composite symbols, as well as any state that might need to be captured in the future.

PiperOrigin-RevId: 289246856
Change-Id: I55baed51f007e3ae6b5d166c55eead1c34f58c4a