Support derived attributes of "list(shape)" kind
Certain ops, like tf.InfeedDequeueTuple, have derived attributes that
are a list of shapes. These derived attributes are based on variadic
operands/results.

This CL adds iterators for querying the shapes of a given variadic
operand/result and registers wrappers in tf_op_base.td for them
so we can define such derived attributes.

Exporting is also updated to handle such derived attributes. Tests
are added. To be consistent, tests for `list(*)` attributes are properly
renamed.

PiperOrigin-RevId: 289526418
Change-Id: I07f60f12436f05e7c1586b5d3ca93774968446a1